add xvfb / notty support

+++ entrypoint.sh
@@ -33,17 +33,22 @@ fi
 # Configure timezone
 ln -snf "/usr/share/zoneinfo/${TZ}" /etc/localtime && echo "${TZ}" > /etc/timezone
 
-# Run in X11 redirection mode (default)
+# Run in X11 redirection mode (default) or with xvfb
 if echo "${RDP_SERVER}" | grep -q -i -E "^(no|off|false|0)$"; then
 
     # Set up pulseaudio for redirection to UNIX socket
     [ -f /root/pulse/client.conf ] && cp /root/pulse/client.conf /etc/pulse/client.conf
 
+    # Run xvfb
+    if echo "${XVFB}" | grep -q -i -E "^(yes|on|true|1)$"; then
+        nohup /usr/bin/Xvfb ${XVFB_SERVER} -screen ${XVFB_SCREEN} ${XVFB_RESOLUTION} >/dev/null 2>&1 &
+    fi
+
     # Run in X11 redirection mode as $USER_NAME (default)
     if echo "${RUN_AS_ROOT}" | grep -q -i -E "^(no|off|false|0)$"; then
 
         # Copy and take ownership of .Xauthority for X11 redirection
-        if [ -f /root/.Xauthority ]; then
+        if [ -f /root/.Xauthority -a "${XVFB}" == "no" ]; then
             cp /root/.Xauthority "${USER_HOME}"
             chown "${USER_UID}":"${USER_GID}" "${USER_HOME}/.Xauthority"
         fi
